Bug fix:修正数据类型与数据库数据类型不匹配问题

+++ src/main/java/common/dao/IProductInfoDAO.java
@@ -4,7 +4,7 @@ import java.math.BigDecimal;
 import java.util.List;
 
 public interface IProductInfoDAO {
-    List<Integer> getAllProductID();
+    List<Long> getAllProductID();
     
-    BigDecimal getProductSalePriceByID(int productID);
+    BigDecimal getProductSalePriceByID(long productID);
 }
